Refactor/move to partial fsd (#24)
* ref: header widget

* fix: react keys

* ref: move css to styles folder

* ref: use module css instead plain css

* feat: header semanticity

* ref: footer widget

* ref: layout component

* feat: page title to shared

* fix: react keys

* ref: useSiteMetadata hook

* ref: move paragraph to shared

* ref: move subtitle to shared

* ref: prices page components
