--- src/components/SiteMetadata.jsx
@@ -1,20 +1,0 @@
-import { graphql, useStaticQuery } from 'gatsby';
-import React from 'react';
-
-const useSiteMetadata = () => {
-  const { site } = useStaticQuery(
-    graphql`
-      query SITE_METADATA_QUERY {
-        site {
-          siteMetadata {
-            title
-            description
-          }
-        }
-      }
-    `,
-  );
-  return site.siteMetadata;
-};
-
-export default useSiteMetadata;
