--- src/components/info/parapgraph/index.js
@@ -1,1 +1,0 @@
-export { Paragraph } from './Paragraph.jsx';
