--- src/components/info/subtitle/index.js
@@ -1,1 +1,0 @@
-export { Subtitle } from './Subtitle.jsx';
