--- src/components/ui/header-burger/HeaderBurger.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import { Body } from './style';
-
-export const HeaderBurger = ({ active, onClick }) => {
-  return (
-    <Body active={active} onClick={onClick}>
-      <span />
-    </Body>
-  );
-};
