--- src/components/ui/header-burger/index.js
@@ -1,1 +1,0 @@
-export { HeaderBurger } from './HeaderBurger.jsx';
