--- src/components/ui/header-form/index.js
@@ -1,1 +1,0 @@
-export { HeaderForm } from './HeaderForm.jsx';
