--- src/components/ui/header-navigation/index.js
@@ -1,1 +1,0 @@
-export { HeaderNavigation } from './HeaderNavigation.jsx';
