--- src/components/ui/header/index.js
@@ -1,1 +1,0 @@
-export { Header } from './Header.jsx';
