--- src/components/ui/page-title/index.js
@@ -1,1 +1,0 @@
-export { PageTitle } from './PageTitle.jsx';
