+++ src/shared/ui/base-page-title/index.js
@@ -1,0 +1,1 @@
+export { BasePageTitle } from './ui/BasePageTitle.jsx';
