+++ src/shared/ui/base-paragraph/index.js
@@ -1,0 +1,1 @@
+export { BaseParagraph } from './ui/BaseParagraph.jsx';
