+++ src/shared/ui/base-subtitle/index.js
@@ -1,0 +1,1 @@
+export { BaseSubtitle } from './ui/BaseSubtitle.jsx';
