+++ src/templates/inspection-page/inspection-page.jsx
@@ -3,23 +3,23 @@ import { graphql } from 'gatsby';
 import React from 'react';
 
 import { Layout } from '../../components/Layout';
-import { Paragraph } from '../../components/info/parapgraph';
-import { Subtitle } from '../../components/info/subtitle';
-import { PageTitle } from '../../components/ui/page-title';
+import { BasePageTitle } from '../../shared/ui/base-page-title';
+import { BaseParagraph } from '../../shared/ui/base-paragraph';
+import { BaseSubtitle } from '../../shared/ui/base-subtitle';
 
 const InspectionPageTemplate = ({ data }) => {
   const { title, subtitle, controlAuthorities } = data.markdownRemark.frontmatter;
 
   return (
     <Layout>
-      <PageTitle text={title} />
-      <Subtitle text={subtitle} />
+      <BasePageTitle text={title} />
+      <BaseSubtitle text={subtitle} />
       {controlAuthorities.map((item, index) => (
-        <Paragraph key={index}>
+        <BaseParagraph key={index}>
           <Typography style={{ fontWeight: 600 }}>{item.name}</Typography>
           <Typography>{item.address}</Typography>
           <Typography>{item.phone}</Typography>
-        </Paragraph>
+        </BaseParagraph>
       ))}
     </Layout>
   );
