+++ src/templates/payment-conditions-page/payment-conditions-page.jsx
@@ -3,31 +3,31 @@ import { graphql } from 'gatsby';
 import React from 'react';
 
 import { Layout } from '../../components/Layout';
-import { Paragraph } from '../../components/info/parapgraph';
-import { Subtitle } from '../../components/info/subtitle';
-import { PageTitle } from '../../components/ui/page-title';
+import { BasePageTitle } from '../../shared/ui/base-page-title';
+import { BaseParagraph } from '../../shared/ui/base-paragraph';
+import { BaseSubtitle } from '../../shared/ui/base-subtitle';
 
 export const PaymentConditionsPageTemplate = ({ props }) => {
   const { title, subtitle, points, payMethods } = props;
 
   return (
     <Layout>
-      <PageTitle text={title} />
-      <Subtitle text={subtitle} />
+      <BasePageTitle text={title} />
+      <BaseSubtitle text={subtitle} />
       {points.map((point, index) => (
-        <Paragraph key={index}>
+        <BaseParagraph key={index}>
           <Typography>
             <strong>{`${index + 1}. ${point.title}`}</strong>
           </Typography>
           <Typography>{point.text}</Typography>
-        </Paragraph>
+        </BaseParagraph>
       ))}
 
-      <Subtitle text="Формы оплаты" />
+      <BaseSubtitle text="Формы оплаты" />
       {payMethods.map((method, index) => (
-        <Paragraph key={index}>
+        <BaseParagraph key={index}>
           <Typography>{`• ${method.text}`}</Typography>
-        </Paragraph>
+        </BaseParagraph>
       ))}
     </Layout>
   );
