+++ src/widgets/footer/ui/Footer.jsx
@@ -1,0 +1,11 @@
+import { Typography } from '@mui/material';
+import { Link } from 'gatsby';
+import React from 'react';
+
+export const Footer = () => {
+  return (
+    <footer>
+      <Typography variant="h3">Footer</Typography>
+    </footer>
+  );
+};
