+++ src/widgets/header/index.js
@@ -1,0 +1,1 @@
+export { Header } from './ui/Header.jsx';
