+++ src/widgets/header/styles/Burger.js
