+++ src/widgets/header/styles/Navigation.js
