+++ src/widgets/header/ui/Burger.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import { Body } from '../styles/Burger';
+
+export const Burger = ({ active, onClick }) => {
+  return (
+    <Body active={active} onClick={onClick}>
+      <span />
+    </Body>
+  );
+};
